Stop processing responses not belonging to a request (#4)
When running as root (or setcap) (network ipX), the listeners started
with each call to `Do` would get a copy of all the incoming responses.
This change allows requests to be sent to multiple hosts concurrently.
Only the request's associated responses are processed.

There can be a further check of ID, but when using udpX pings
(`net.ipv4.ping_group_range` set), the ID in the response is always a
completely different number and causes timeouts. It is unclear why this
number is so off.